feat(waybar): add if else to launch_me script

+++ desktop/waybar/scripts/launch_me.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 
-pkill waybar
-waybar -c ~/.config/waybar/config.json &
+if pgrep -x waybar > /dev/null
+then
+	pkill waybar
+else
+	echo "Starting Waybar..."
+	waybar -c ~/.config/waybar/config.json &
+fi
